--- a/test/CMakeLists.txt
+++ b/test/CMakeLists.txt
@@ -1,8 +1,11 @@
+add_library(catch STATIC test_main.cpp)
+
 macro(_datadog_test TEST_NAME)
   add_executable(${TEST_NAME} ${ARGN})
   add_sanitizers(${TEST_NAME})  
   target_link_libraries(${TEST_NAME} ${DATADOG_LINK_LIBRARIES}
-                                     dd_opentracing)
+                                     dd_opentracing
+                                     catch)
   add_test(${TEST_NAME} ${TEST_NAME})
 endmacro()