--- a/test/agent_writer_test.cpp
+++ b/test/agent_writer_test.cpp
@@ -5,7 +5,6 @@
 
 #include <ctime>
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;