--- a/test/opentracing_test.cpp
+++ b/test/opentracing_test.cpp
@@ -1,7 +1,6 @@
 #include <datadog/opentracing.h>
 #include "mocks.h"
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;