--- a/test/propagation_test.cpp
+++ b/test/propagation_test.cpp
@@ -7,7 +7,6 @@
 #include "../src/tracer.h"
 #include "mocks.h"
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;
 namespace tags = datadog::tags;