--- a/test/sample_test.cpp
+++ b/test/sample_test.cpp
@@ -6,7 +6,6 @@
 
 #include <ctime>
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;