--- a/test/span_buffer_test.cpp
+++ b/test/span_buffer_test.cpp
@@ -2,7 +2,6 @@
 #include "../src/sample.h"
 #include "mocks.h"
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;