--- a/test/span_test.cpp
+++ b/test/span_test.cpp
@@ -7,7 +7,6 @@
 #include "../src/sample.h"
 #include "mocks.h"
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;
 namespace tags = datadog::tags;