--- a/test/tracer_factory_test.cpp
+++ b/test/tracer_factory_test.cpp
@@ -4,7 +4,6 @@
 // Source file needed to ensure compilation of templated class TracerFactory<MockTracer>
 #include "../src/tracer_factory.cpp"
 
-#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 using namespace datadog::opentracing;