This PR doesn't actually enhance the library performance, but compile time when building with `BUILD_TESTING=ON` is faster by around 50%.

I've measured compilation time with `time (cmake -DBUILD_TESTING=ON .. && make && ctest --output-on-failure)`.
Before:
`128,05s user 5,15s system 98% cpu 2:14,64 total`
After:
`72,01s user 3,73s system 98% cpu 1:17,00 total`

As recommended by Catch2 documentation on this [README](https://github.com/catchorg/Catch2/blob/master/docs/slow-compiles.md) it's better to build Catch2 `main` in a separate static library and link each test to it, this way Catch2 `main` is generated and built only once and changing a test doesn't require rebuilding it.